+++ Library/PTPusher.h
@@ -94,19 +94,6 @@ extern NSString *const PTPusherErrorUnderlyingEventKey;
 @property (nonatomic, unsafe_unretained) id<PTPusherDelegate> delegate;
 #endif
 
-
-
-/** This property is deprecated and will be ignored.
- *
- * The client will always attempt to reconnect when it disconnects as long as the error 
- * code it disconnected with permits automatic reconnection (see the Pusher protocol documentation
- * for an overview of error codes: http://pusher.com/docs/pusher_protocol).
- *
- * If you want to prevent the client from automatically reconnecting, you can do so by returning
- * NO from the delegate method pusher:connectionWillConnect:.
- */
-@property (nonatomic, assign, getter=shouldReconnectAutomatically) BOOL reconnectAutomatically __PUSHER_DEPRECATED__;
-
 /** Specifies the delay between reconnection attempts. Defaults to 5 seconds.
  */
 @property (nonatomic, assign) NSTimeInterval reconnectDelay;